+++ R/simulation_settings.R
@@ -10,6 +10,7 @@
 #' @slot nocb NOCB settings object
 #' @slot declare declare settings (mrgsolve only)
 #' @slot progress progress settings
+#' @slot replication replication settings
 #' @slot internal internal settings
 #' @export
 setClass(
@@ -20,15 +21,17 @@ setClass(
     nocb="nocb_settings",
     declare="declare_settings",
     progress="progress_settings",
+    replication="replication_settings",
     internal="internal_settings"
   ),
-  prototype=prototype(hardware=Hardware(), solver=Solver(), nocb=NOCB(), declare=Declare(), progress=Progress())
+  prototype=prototype(hardware=Hardware(), solver=Solver(), nocb=NOCB(), declare=Declare(),
+                      progress=Progress(), replication=AutoReplicationSettings())
 )
 
 #'
 #' Create advanced simulation settings.
 #'
-#' @param ... any user-required settings: see ?Hardware, ?Solver, ?NOCB, ?Declare or ?Progress settings
+#' @param ... any user-required settings: see ?Hardware, ?Solver, ?NOCB, ?Declare, ?Progress or ?AutoReplicationSettings
 #' @return advanced simulation settings
 #' @importFrom purrr detect
 #' @export
@@ -65,17 +68,24 @@ Settings <- function(...) {
     progress <- Progress()
   }
   
+  # Check if replication settings are specified
+  replication <- args %>% purrr::detect(~(is(.x, "replication_settings")))
+  if (is.null(replication)) {
+    replication <- AutoReplicationSettings()
+  }
+  
   # Check no other argument remains
   others <- args %>%  purrr::discard(~(is(.x, "hardware_settings") ||
                                        is(.x, "solver_settings") ||
                                        is(.x, "nocb_settings") ||
                                        is(.x, "declare_settings") ||
-                                       is(.x, "progress_settings")
+                                       is(.x, "progress_settings") ||
+                                       is(.x, "replication_settings")  
                                        ))
   assertthat::assert_that(length(others) == 0,
-                          msg="Unknown argument detected. Accepted settings: see ?Hardware, ?Solver, ?NOCB, ?Declare or ?Progress")
+                          msg="Unknown argument detected. Accepted settings: see ?Hardware, ?Solver, ?NOCB, ?Declare, ?Progress or ?AutoReplicationSettings")
   
-  return(new("simulation_settings", hardware=hardware, solver=solver, nocb=nocb, declare=declare, progress=progress))
+  return(new("simulation_settings", hardware=hardware, solver=solver, nocb=nocb, declare=declare, progress=progress, replication=replication))
 }
 
 #_______________________________________________________________________________
@@ -89,4 +99,5 @@ setMethod("show", signature=c("simulation_settings"), definition=function(object
   show(object@nocb)
   show(object@declare)
   show(object@progress)
+  show(object@replication)
 })